+++ apps/sim/socket-server/handlers/subblocks.ts
@@ -158,147 +158,4 @@ export function setupSubblocksHandlers(
       })
     }
   })
-
-  socket.on('variable-update', async (data) => {
-    const workflowId = roomManager.getWorkflowIdForSocket(socket.id)
-    const session = roomManager.getUserSession(socket.id)
-
-    if (!workflowId || !session) {
-      logger.debug(`Ignoring variable update: socket not connected to any workflow room`, {
-        socketId: socket.id,
-        hasWorkflowId: !!workflowId,
-        hasSession: !!session,
-      })
-      return
-    }
-
-    const { variableId, field, value, timestamp, operationId } = data
-    const room = roomManager.getWorkflowRoom(workflowId)
-
-    if (!room) {
-      logger.debug(`Ignoring variable update: workflow room not found`, {
-        socketId: socket.id,
-        workflowId,
-        variableId,
-        field,
-      })
-      return
-    }
-
-    try {
-      const userPresence = room.users.get(socket.id)
-      if (userPresence) {
-        userPresence.lastActivity = Date.now()
-      }
-
-      // First, verify that the workflow still exists in the database
-      const workflowExists = await db
-        .select({ id: workflow.id })
-        .from(workflow)
-        .where(eq(workflow.id, workflowId))
-        .limit(1)
-
-      if (workflowExists.length === 0) {
-        logger.warn(`Ignoring variable update: workflow ${workflowId} no longer exists`, {
-          socketId: socket.id,
-          variableId,
-          field,
-        })
-        roomManager.cleanupUserFromRoom(socket.id, workflowId)
-        return
-      }
-
-      let updateSuccessful = false
-      await db.transaction(async (tx) => {
-        const [workflowRecord] = await tx
-          .select({ variables: workflow.variables })
-          .from(workflow)
-          .where(eq(workflow.id, workflowId))
-          .limit(1)
-
-        if (!workflowRecord) {
-          // Workflow was deleted - this is a normal race condition in collaborative editing
-          logger.debug(
-            `Ignoring variable update for deleted workflow: ${workflowId}/${variableId}.${field}`
-          )
-          return
-        }
-
-        const variables = (workflowRecord.variables as any) || {}
-
-        if (!variables[variableId]) {
-          // Variable was deleted - this is a normal race condition in collaborative editing
-          logger.debug(
-            `Ignoring variable update for deleted variable: ${workflowId}/${variableId}.${field}`
-          )
-          return
-        }
-
-        // Update the variable field
-        variables[variableId] = {
-          ...variables[variableId],
-          [field]: value,
-        }
-
-        await tx
-          .update(workflow)
-          .set({
-            variables: variables,
-            updatedAt: new Date(),
-          })
-          .where(eq(workflow.id, workflowId))
-
-        updateSuccessful = true
-      })
-
-      // Only broadcast to other clients if the update was successful
-      if (updateSuccessful) {
-        socket.to(workflowId).emit('variable-update', {
-          variableId,
-          field,
-          value,
-          timestamp,
-          senderId: socket.id,
-          userId: session.userId,
-        })
-
-        // Emit confirmation if operationId is provided
-        if (operationId) {
-          socket.emit('operation-confirmed', {
-            operationId,
-            serverTimestamp: Date.now(),
-          })
-        }
-
-        logger.debug(`Variable update in workflow ${workflowId}: ${variableId}.${field}`)
-      } else if (operationId) {
-        // Variable was deleted - notify client that operation completed (but didn't update anything)
-        socket.emit('operation-failed', {
-          operationId,
-          error: 'Variable no longer exists',
-          retryable: false, // No point retrying for deleted variables
-        })
-      }
-    } catch (error) {
-      logger.error('Error handling variable update:', error)
-
-      const errorMessage = error instanceof Error ? error.message : 'Unknown error'
-
-      if (operationId) {
-        socket.emit('operation-failed', {
-          operationId,
-          error: errorMessage,
-          retryable: true,
-        })
-      }
-
-      // Legacy operation-error for backward compatibility
-      socket.emit('operation-error', {
-        type: 'VARIABLE_UPDATE_FAILED',
-        message: `Failed to update variable ${variableId}.${field}: ${errorMessage}`,
-        operation: 'variable-update',
-        target: 'variable',
-      })
-    }
-  })
 }
